Add files via upload
n this lecture, we will discuss the topic of morphology, which refers to the study of the form and structure of objects. In the context of image processing, morphological operations are used to process segmented images by removing imperfections and enhancing the structural information.

There are two types of morphological operations: simple and compound. Simple morphological operations involve basic operations such as erosion, dilation, opening, and closing. Erosion is used to erode the boundaries of objects in an image, while dilation is used to expand the boundaries of objects. Opening is a combination of erosion followed by dilation, and it is used to remove small noise or imperfections in an image. Closing is a combination of dilation followed by erosion, and it is used to fill gaps or holes in objects.

Compound morphological operations involve more complex operations, such as top-hat and bottom-hat operations. Top-hat operation is the difference between the original image and its opening, and it is used to enhance the brighter regions of an image. Bottom-hat operation is the difference between the closing of an image and the original image, and it is used to enhance the darker regions of an image.